fix(distance): fix distance map computation and improve distance calculation to use uint32 and handle distance sum stored in uint64

+++ pkg/common/types.go
@@ -63,7 +63,7 @@ const (
 	TIMESTAMP_DEPENDENCY_ORACLE OracleType = "timestamp-dependency"
 )
 
-type DistanceMap map[string]map[string]uint64 // blockPC => instruction => distance
+type DistanceMap map[string]map[string]uint32 // blockPC => instruction => distance
 
 type PowerScheduleStrategy string
